requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
boto3==1.34.21
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.10.1
bcrypt==5.0.0
